requirements.txt
openai
python-dotenv
apscheduler
requests
colorama
PyYAML
flask
python-docx
wechatpy
cryptography
